Add comments helper (#1007)
* Create comments_helper.rb

- Methods all copied from application_controller.rb

* Update application_controller.rb to use helper

* Update meetings_controller.rb to use helper

* Update moments_controller.rb to use helper

* Update strategies_controller.rb to use helper

* Add comment to make Rubocop happy

* Move comment-related tests to new helper spec
