+++ app/controllers/moments_controller.rb
@@ -3,6 +3,7 @@
 # rubocop:disable ClassLength
 class MomentsController < ApplicationController
   include CollectionPageSetup
+  include CommentsHelper
   include Shared
 
   before_action :set_moment, only: %i[show edit update destroy]
